Update to Python/Linux MAC changer
Have removed all shell=True parameters because it is a security hazard. Removed spacing and added in the subprocess calls to a list instead to prevent misuse of the application. Prior to this change, anyone could have hijacked the application and used it to run commands openly instead of those specified.